+++ graph_email.py
@@ -1,4 +1,3 @@
-from typing import List
 from azure.identity.aio import DefaultAzureCredential
 from msgraph import GraphServiceClient
 from msgraph.generated.models.email_address import EmailAddress
@@ -8,23 +7,27 @@ from msgraph.generated.models.message import Message
 from msgraph.generated.models.recipient import Recipient
 from msgraph.generated.users.item.send_mail.send_mail_post_request_body import SendMailPostRequestBody
 
+from report_utils import Email
+
 MSGRAPH_SCOPES = ["https://graph.microsoft.com/.default"]
 
 
-async def send_email(subject: str, body: str, to_addresses: List[str]):
+async def send_email(email: Email):
     async with DefaultAzureCredential() as credential:
         client = GraphServiceClient(credentials=credential, scopes=MSGRAPH_SCOPES)
 
-    to_recipients = [Recipient(email_address=EmailAddress(address=a)) for a in to_addresses]
+    to_recipients = [Recipient(email_address=EmailAddress(address=a)) for a in email.to]
+    cc_recipients = [Recipient(email_address=EmailAddress(address=a)) for a in email.cc]
 
     request_body = SendMailPostRequestBody(
         message=Message(
-            subject=subject,
+            subject=email.subject,
             body=ItemBody(
                 content_type=BodyType.Html,
-                content=body,
+                content=email.body,
             ),
             to_recipients=to_recipients,
+            cc_recipients=cc_recipients,
         ),
     )
